refactor(auto-reply): isolate session conflict retry
