+++ src/auto-reply/reply/session.ts
@@ -53,7 +53,6 @@ import {
   forgetActiveSessionForShutdown,
   noteActiveSessionForShutdown,
 } from "../../gateway/active-sessions-shutdown-tracker.js";
-import { computeBackoff, sleepWithAbort, type BackoffPolicy } from "../../infra/backoff.js";
 import { getSessionBindingService } from "../../infra/outbound/session-binding-service.js";
 import { deliverSessionMaintenanceWarning } from "../../infra/session-maintenance-warning.js";
 import { createSubsystemLogger } from "../../logging/subsystem.js";
@@ -99,6 +98,10 @@ import {
   type ReplySessionEntryHandle,
 } from "./session-entry-handle.js";
 import { buildSessionEndHookPayload, buildSessionStartHookPayload } from "./session-hooks.js";
+import {
+  ReplySessionInitConflictError,
+  runWithSessionInitConflictRetry,
+} from "./session-init-conflict-retry.js";
 import { prepareReplySessionParentFork } from "./session-parent-fork-prepare.js";
 import { clearSessionResetRuntimeState } from "./session-reset-cleanup.js";
 import {
@@ -333,68 +336,6 @@ export function resolveReplySessionPreprocessingState(
     sessionKey,
     storePath: attemptContext.storePath,
   };
-}
-
-/**
- * Raised when the optimistic-concurrency commit of reply session
- * initialization loses the revision compare-and-swap. The message shape is
- * load-bearing: channel monitors (e.g. the Telegram poller) match it to
- * classify the failure as retryable.
- */
-export class ReplySessionInitConflictError extends Error {
-  constructor(sessionKey: string) {
-    super(`reply session initialization conflicted for ${sessionKey}`);
-    this.name = "ReplySessionInitConflictError";
-  }
-}
-
-const SESSION_INIT_CONFLICT_MAX_ATTEMPTS = 5;
-const SESSION_INIT_CONFLICT_BACKOFF_POLICY = {
-  initialMs: 250,
-  maxMs: 4_000,
-  factor: 2,
-  jitter: 0.05,
-} satisfies BackoffPolicy;
-
-/**
- * Retries `attempt` on session-init CAS conflicts with jittered exponential
- * backoff. Conflicts are expected when another writer (restart recovery,
- * a concurrent dashboard tab, an active reply run) is mid-commit on the same
- * session key; those commits settle in well under a second, so a delayed
- * fresh attempt almost always wins. Must only wrap the *unlocked* outer
- * attempt: sleeping while holding the session-store writer lane would stall
- * every other writer on the store.
- */
-export async function runWithSessionInitConflictRetry<T>(
-  attempt: () => Promise<T>,
-  options?: {
-    maxAttempts?: number;
-    signal?: AbortSignal;
-    sleep?: (ms: number, signal?: AbortSignal) => Promise<void>;
-  },
-): Promise<T> {
-  const maxAttempts = options?.maxAttempts ?? SESSION_INIT_CONFLICT_MAX_ATTEMPTS;
-  const sleep = options?.sleep ?? sleepWithAbort;
-  for (let attemptIndex = 0; ; attemptIndex += 1) {
-    try {
-      return await attempt();
-    } catch (error) {
-      if (
-        !(error instanceof ReplySessionInitConflictError) ||
-        attemptIndex >= maxAttempts - 1 ||
-        options?.signal?.aborted === true
-      ) {
-        throw error;
-      }
-      const backoffMs = computeBackoff(SESSION_INIT_CONFLICT_BACKOFF_POLICY, attemptIndex + 1);
-      log.debug(
-        `reply session initialization conflicted; retrying in ${backoffMs}ms (attempt ${attemptIndex + 2}/${maxAttempts})`,
-      );
-      // Cancellation must interrupt the wait itself; otherwise shutdown can
-      // sleep through the backoff and start one more session-init attempt.
-      await sleep(backoffMs, options?.signal);
-    }
-  }
 }
 
 /** Initializes or reuses the reply session state for one inbound turn. */
